refactor: Separate Create component from App component
